+++ .reaction/pluginConfig.js
